Calculate the resource usage of the shoot control planes (!)
The cadvisor metrics about the CPU, Memory and Network usage:

- container_cpu_usage_seconds_total

- container_memory_usage_bytes
- container_memory_working_set_bytes

- container_network_receive_bytes_total
- container_network_transmit_bytes_total

the kube-state-metrics metrics about the CPU, Memory and Disk requests:

- kube_pod_container_resource_requests (for both CPU and memory)
- kube_persistentvolumeclaim_resource_requests_storage_bytes

and the kubelet metrics about the Disk usage:

- kubelet_volume_stats_used_bytes

are aggregated

- by namespace
- by namespace and owner (and container, if applicable).

Instead of pod names that contain random suffixes, the name of the respective
*owner* resource (e.g. deployment, statefulset, daemonset) is used for the
aggregation. This might help to significantly reduce the cardinality of the
metrics and hence the load on Prometheus when these metrics are stored for a
longer period of time.

The owner and container labels might help to find the components that contribute
most to the CPU or Memory usage. The owner label might help to find the
components that contribute most to the Network or Disk usage. (The container
label is not available for the Network and Disk metrics.)

The name of the namespace that hosts the shoot control plane is derived from the
name of the project and the shoot, but it does not uniquely identify the shoot.
The shoot's *UID* is included as an additional label to allow to associate the
information with a specific shoot.

*Date* (year, month) labels are added to allow for calculating the average usage
for each month.

  To facilitate testing this feature e.g. in the "remote local setup", the date
  labels also include the *day* label besides the year and month labels so that
  the transition to the next period happens every day at UTC midnight, not only
  each month. To reproduce most of the edge cases, I suggest creating a
  permanently active, a regularly hibernated shoot cluster and regularly
  creating and deleting another shoot cluster with the same name to simulate the
  churn.

    while true; do
      sleep 30m
      kubectl patch shoot -n garden-local local3 -p '{"spec":{"hibernation":{"enabled": true}}}'
      sleep 30m
      kubectl patch shoot -n garden-local local3 -p '{"spec":{"hibernation":{"enabled": false}}}'
    done

    while true; do
      kubectl apply -f example/provider-local/shoot.yaml
      kubectl wait --for=condition=ControlPlaneHealthy shoot local --timeout=-1s
      sleep 5m
      kubectl annotate shoot local confirmation.gardener.cloud/deletion=true
      kubectl delete shoot local
      sleep 5m
    done

The promql expressions are somewhat more advanced, please refer to the official
documentation:

- https://prometheus.io/docs/prometheus/latest/querying/basics/
- https://prometheus.io/docs/prometheus/latest/querying/operators/
- https://prometheus.io/docs/prometheus/latest/querying/functions/

The first few recording rules that are prefixed with "_" were introduced to
improve the readability of the implementation.

_namespace_to_shoot_uid uses the label_replace function to give a shorter name
to the label holding the shoot UID: `shoot_uid` instead of
`annotation_shoot_gardener_cloud_uid`.

_namespace_to_shoot_uid_and_date adds the year and month labels. It uses the
timestamp and count_values functions to achieve that.

The idiom to "join" two instant vectors on a set of labels to add some other
labels to the resulting vector is used extensively to enrich the raw metrics
with additional metadata: the owner name (instead of the pod name), the shoot
UID and the date.

The following metrics are calculated:

CPU, Memory, Network and Disk usage and requests by namespace. These metrics
carry only the namespace, shoot_uid and date (year, month) labels so that they
have a relatively low cardinality and could be used for long term storage.

- metering  :cpu_usage                 :sum_by_namespace
- metering  :cpu_requests              :sum_by_namespace

- metering  :memory_usage              :sum_by_namespace
- metering  :working_set_memory        :sum_by_namespace
- metering  :memory_requests           :sum_by_namespace

- metering  :network_transmit          :sum_by_namespace
- metering  :network_receive           :sum_by_namespace

- metering  :persistent_volume_claims  :sum_by_namespace
- metering  :persistent_volume_usage   :sum_by_namespace

CPU, Memory, Network and Disk usage and requests by namespace, owner and
container. These metrics carry in addition the owner and container labels and
hence can have a significantly higher (100x) cardinality than the previous
metrics. They might be required to find the culprit of high resource usage but
it might not be necessary to store them indefinitely.

- metering  :cpu_usage                 :sum_by_namespace_owner_container
- metering  :cpu_requests              :sum_by_namespace_owner_container

- metering  :memory_usage              :sum_by_namespace_owner_container
- metering  :working_set_memory        :sum_by_namespace_owner_container
- metering  :memory_requests           :sum_by_namespace_owner_container

- metering  :network_transmit          :sum_by_namespace_owner
- metering  :network_receive           :sum_by_namespace_owner

- metering  :persistent_volume_claims  :sum_by_namespace_owner
- metering  :persistent_volume_usage   :sum_by_namespace_owner

The promQL expressions are canonically formatted in the Prometheus UI.

The calculated metrics can be federated from all the seed clusters to get a
landscape wide overview of the shoot cluster control plane resource usage.

The `/rules#metering.rules` endpoint of the Prometheus UI can be used to monitor
the total evaluation time (i.e. CPU usage) of all the recording rules above.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>