Allow list the kube_{pod,replicaset}_owner metric for the cache Prometheus
These metrics are useful to follow the owner reference from the pod to the
deployment (via the replicaset) and to the daemonset or statefulset resources.

Deployments, Statefulsets and Daemonsets generate Pods with names
that have one or two random suffixes.

By following the owner references we can map the partially random pod
names back to the name of their owners without the random suffix.

Using the owner label instead of the partially random pod label can
help to reduce the cardinality related load on Prometheus.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>