Federate the metering metrics to the aggregate Prometheus
The aggregate Prometheus is configured as a datasource for the garden/grafana
deployment so that dashboards can visualize the metering metrics there. It can
also be used to federate the metering metrics from the seeds via an ingress of
the aggregate Prometheus.

Note that only the "raw", non stateful metering metrics are federated.

The stateful metering metrics are recalculated independently in the aggregate
Prometheus.

Recalculating these stateful metrics on all the layers allow to add redundancy
in the lower layers, e.g. the aggregate Prometheus could scrape 2 cache
Prometheus instances via a service, for higher availability.

This also saves network traffic at the price of a bit more CPU usage.

Note that the stateful recording rules are symlinked to avoid code duplication.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>