Allow list the persistent volume usage metric for the cache Prometheus
This metric is exposed by the kubelet on the /metrics endpoint. Not to be
confused with the cadvisor metrics that are also exposed by the kubelet on the
/metrics/cadvisor endpoint.

```
k proxy
curl -s http://localhost:8001/api/v1/nodes/$(k get nodes -o json | jq '.items[0].metadata.name' -r)/proxy/metrics | less -SNM
```

Note that this metric is not exposed by the kubelet in the "local setup"
probably because the persistent volumes are mapped to a local path by the
"local-path-provisioner" component. This means that the disk usage metrics are
going to show as 0% in the local setup.

Note that this scrape configuration is deployed via a helm chart, unlike the
previous commits which manipulate the allow list for kube-state-metrics that is
deployed via a Go component.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>