Let prometheus ignore the exposed timestamps when scraping metrics
The cadvisor exposes timestamps with its metrics:

  k proxy &
  curl -s http://localhost:8001/api/v1/nodes/$(k get nodes -o json | jq '.items[0].metadata.name' -r)/proxy/metrics/cadvisor | head
  ...
  container_blkio_device_usage_total{container="",device="/dev/sda",id="/",image="",major="8",minor="0",name="",namespace="",operation="Async",pod=""} 5.245841408e+09 1667569445770

and also Prometheus exposes timestamps when it is federated:

  curl -k -G --data-urlencode 'match[]={__name__=~"metering:.+"}' localhost:9090/federate | head
  ...
  metering:cpu_requests:sum_by_namespace{month="11",namespace="extension-networking-calico-ds88k",year="2022",instance=""} 0.01 1667569547691

According to the document
https://promcon.io/2017-munich/slides/staleness-in-prometheus-2-0.pdf
the new staleness handling in Prometheus 2.0 does not work if timestamps are
exposed and honored.

With this change we let Prometheus ignore the timestamps so that series that are
missing in a scrape are marked as stale right away.

This helps to avoid overlapping time series for 5 minutes and artifacts in
recording rules.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>