Keep 1 replica of loki, grafana-{users,operators} for a transition period
After replacing Loki with Vali, new logs are going to be stored in Vali
and the older logs in Loki are not going to be migrated to Vali.

To allow Gardener operators access to the older logs in Loki,
the former Loki (and Grafana) instances are going to be kept around
in the shoot control plane namespace,
at least for the default log retention period of two weeks.

TODO check and document how to access the older logs in Loki.

This commit adds a simple logic to the logging reconcile function
to keep 1 replica of loki, grafana-users and grafana-operators.

Co-authored-by: Niki Dokovski <nickytd@gmail.com>
Co-authored-by: Kristian-ZH <k.zhelyazkov@sap.com>
Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>