Temporarily update the gardener/logging images to a dev version
The PR https://github.com/gardener/logging/pull/166 in the
gardener/logging repository replaces Loki with Vali, however,
we plan to merge/release that PR only when this gardener PR is about to be
ready, to allow for creating new gardener/logging releases before
the Loki -> Vali replacement (if needed).

This commit updates the 4 images to a development version of the
gardener/logging repository that was built with the Loki -> Vali PR above:

  https://github.com/gardener/logging/blob/096cadfb857636f7fbc01e1d66277b9afc6b2f71/.ci/pipeline_definitions#L14-L34

  - gardener/fluent-bit-to-vali
  - gardener/vali-curator
  - gardener/telegraf-iptables
  - gardener/event-logger

This allows to test further aspects of this PR in a local setup.

Co-authored-by: Niki Dokovski <nickytd@gmail.com>
Co-authored-by: Kristian-ZH <k.zhelyazkov@sap.com>
Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>