Adjust the Vali prefix (and keep the Plutono prefix)
The Vali prefix is used to ship logs from the shoot nodes to the control plane
and hence it can be changed consistently during the shoot reconciliation.

The Plutono prefix is deliberately not changed here and will be adjusted in a
future PR. Currently the Grafana/Plutono prefix (gu) is also hard coded in the
Dashboard and to avoid the need for a coordinated release with the Dashboard and
to avoid showing an incorrect link in the Dashboard until the shoot cluster is
reconciled, we rather postpone this cleanup to a future PR where maybe the
Dashboard will be able to pick the Plutono URL from the monitoring secret and
hence will support a smooth transition to a new prefix.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>