Delete Loki (except for its PVC+PV)
During the migration to Vali, we delete the Loki StatefulSet
and related resources. They are going to be replaced by the Vali artefacts.
The exception here is the PersistentVolumeClaim, which can be reused with
Vali, thereby preserving the logs stored in Loki.

The gardenlet needs to be able to delete the old loki artefacts.

The newly added functions are copies of their loki counterparts.

This commit can be reverted in a future release of Gardener.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>