Improve health check during transitional period from Loki to Vali
In the time period between a seed being upgraded to the release that
contains the Vali changes and the shoot being reconciled, the seed's
gardenlet should not expect the Vali StatefulSet to be healthy, but
instead expect the old Loki StatefulSet to be healthy.

This commit can be reverted in a future release.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>