Improve health check during the transition period
In the time period between a seed being upgraded to the release that contains
this PR and the shoot being reconciled, the seed's gardenlet should not expect
the Plutono/Vali artifacts to be healthy, but instead expect the old
Grafana/Loki deployment or statefulset to be there.

This commit can be reverted in a future release.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>