Improve health check during transitional period from Grafana to Plutono
In the time period between a seed being upgraded to the release that
contains the Plutono changes and the shoot being reconciled, the seed's
gardenlet should not expect the Plutono deployment to be healthy, but
instead expect the old Grafana deployment to be healthy.

This commit can be reverted in a future release.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>