Fix data source names for the extension dashboards
If an extension contributes dashboards to Grafana/Plutono, their annotation
data source may need to be adjusted to the Grafana/Plutono transition,
because it may contain the hardcoded string `Grafana`, which needs to be
replaced with `Plutono`. Dashboards could also be using a hardcoded string
`loki` as a data source for showing logs, which needs to be renamed to
`vali`. We replace these strings in memory. This is to ensure that when we
transition to Plutono/Vali, extension dashboards can be loaded successfully
in Plutono. Later, and without time pressure, we can adapt the dashboards
in the individual extension repositories.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>