Delete Loki (except for its PVC+PV)
During the migration from Loki to Vali, we delete the Loki StatefulSet and
related resources. They are replaced by the Vali artefacts.

The exception here is the PersistentVolumeClaim, which can be reused with Vali,
thereby preserving the logs stored in Loki.

The gardenlet needs to be able to delete the old loki artefacts, hence the
cluster role is adjusted.

The newly added functions are copies of their "Vali" counterparts.

The deletion is guarded by the lokiPvcExists condition: the existence of a
loki-loki-0 persistent volume claim indicates that there is a loki based logging
stack and it shall be deleted. The existing unit tests are adjusted so that the
mock client returns false for the lokiPvcExists condition.

This commit can be reverted in a future release of Gardener.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>