Rename the Loki PVC to Vali (!)
The name of a PVC is derived from the

- the name of the pod which is created by the StatefulSet, e.g. vali-0, and
- the name of the volume claim template in the StatefulSet, e.g. vali.

So the `loki` StatefulSet uses a `loki-loki-0` PVC, whereas the new `vali`
StatefulSet uses a `vali-vali-0` PVC.

The StatefulSet creates a new PVC based on the PVC template if one does not
exist, or it uses the existing one with the matching name.

The intention of this PR is to reuse the disk of Loki for Vali so that the 2
weeks worth of logs that were ingested into Loki are not lost due to the
migration to Vali.

Hence, to reuse the disk of the Loki StatefulSet for the Vali StatefulSet, we
need to rename the `loki-loki-0` PVC to `vali-vali-0`.

It is unfortunately not possible to "simply" rename a PVC: the name is part of
the identity of a Kubernetes resource and it can not be changed. However, with
the approach described in
https://github.com/gardener/gardener/pull/7318#issuecomment-1460081163 it is
possible to create a new PVC with the new name `vali-vali-0` that is bound to
the persistent volume of the `loki-loki-0` PVC.

The trick is to change the PV's ReclaimPolicy temporarily to Retain and delete
the Loki PVC. The PV will not be deleted due to the Retain reclaim policy. After
setting the claimRef of the PV to nil, a new Vali PVC can be created which will
be bound by the PV/PVC controller in the kube controller manager to the existing
PV. Ultimately this leads to a state that is similar to "simply" renaming the
PVC.

An executable example flow to rename loki to vali in the control plane looks
like this:

    #!/bin/bash -exv

    # Mark the volume of loki
    kubectl exec loki-0 -c loki -- sh -c "echo loki > /data/hello"

    kubectl get sts loki -o yaml        | kubectl neat > sts.yml
    kubectl get pvc loki-loki-0 -o yaml | kubectl neat > pvc.yml

    kubectl delete sts loki

    pv_id=$(kubectl get pvc loki-loki-0 -o jsonpath='{.spec.volumeName}')
    kubectl patch pv "$pv_id" -p '{"spec":{"persistentVolumeReclaimPolicy":"Retain"}}'

    kubectl delete pvc loki-loki-0
    kubectl patch pv "$pv_id" -p '{"spec":{"claimRef": null }}'

    sed 's/loki-loki-0/vali-vali-0/g' pvc.yml | kubectl apply -f -
    sed 's/name: loki$/name: vali/g'  sts.yml | kubectl apply -f -

    kubectl patch pv "$pv_id" -p '{"spec":{"persistentVolumeReclaimPolicy":"Delete"}}'

    # Wait for the vali StatefulSet to be ready
    kubectl rollout status sts/vali

    # Verify that the volume of vali is the former loki volume
    kubectl exec vali-0 -c vali -- cat /data/hello

Hopefully the explanation above helps to follow the coding that is added in this
commit.

In case of unexpected errors, we accept that we can not reuse the Loki PV and
attempt to clean up the PV/PVC so that the next reconciliation can proceed with
ultimately creating a new PV for Vali.

Note that after we successfully deleted the Loki PVC, we can't easily find in
the next iteration the corresponding (unnamespaced) Loki PV, hence we try to
do the clean up in the first iteration.

One expected error cases is when the 30s deadline of the shoot reconciliation is
reached, so we use a new context with a 1 minute deadline for the more critical
steps in the second half of the flow and also for the recovery steps after an
unexpected error.

Note that the loki-0 pod currently can not terminate gracefully in a timely
manner: it is eventually killed by the kubelet after 30s. We reduce the graceful
termination timeout to 5 seconds when deleting Loki because based on its logs,
loki is itself waiting for 30s after an almost complete graceful shutdown in 1s.
By reducing the graceful termination timeout we can use the remaining time to
proceed with the flow above.

This commit can be reverted in a future release of Gardener.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>