WIP! Allow the fluentbits to talk to loki during the transition phase
When the gardenlet is updated to this PR, but the shoots are not yet
reconciled, the fluentbits in the garden namespace are sending
the control plane logs to the logging service in the control plane
which still points to the loki-0 pod.

By adding this label, the generated network policies allow the network
communication between the fluentbits and loki-0.

When the shoot is reconciled, loki-0 is going to be replaced by vali-0,
and that case is already covered by the network policy label in the line
above.

This commit can be reverted in a future release of Gardener.

Co-authored-by: Niki Dokovski <nickytd@gmail.com>
Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>

Conflicts:
  charts/seed-bootstrap/charts/fluent-bit/templates/fluent-bit-daemonset.yaml

LOCAL: deleted
REMOTE:
@@ -24,10 +24,11 @@ spec:
 {{ toYaml .Values.labels | indent 8 }}
         networking.gardener.cloud/to-dns: allowed
         networking.gardener.cloud/to-runtime-apiserver: allowed
         networking.resources.gardener.cloud/to-logging-tcp-3100: allowed
         networking.resources.gardener.cloud/to-all-shoots-vali-tcp-3100: allowed
+        networking.resources.gardener.cloud/to-all-shoots-loki-tcp-3100: allowed