Add the hash of the labels as a suffix to the fluentbit custom resource
The fluentbit custom resource contains a single set of labels, that is used for

- labels of the daemon set
- label selector of the daemon set
- labels of the pod template

Generally it is possible to change the labels of the pod template, but the label
selector field of a daemonset is immutable. The fluent operator uses the same
set of labels for both, which means that we can not change the pod template
labels.

A mitigation is to add a hash of the labels to the name of the fluentbit custom
resource. When we change the labels, we'll get a new hash. The gardener resource
manager will delete the old fluentbit resource and create a new one. The
fluentbit operator will delete the old daemon set and create a new one. Hence we
won't face the issue of trying to change an immutable field in the daemonset.

Co-authored-by: Niki Dokovski <nickytd@gmail.com>
Co-authored-by: Kristian-ZH <k.zhelyazkov@sap.com>
Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>