Configure Cache Prometheus to scrape kubelet and cadvisor metrics directly
Previously, the Cache Prometheus scraped the kubelet and cadvisor metrics from
the seed nodes via the Kubernetes API server's proxy feature. This is
unnecessary as the Cache Prometheus can directly access the cluster nodes.

This change enhances the Cache Prometheus configuration to scrape metrics
directly from the cluster nodes, thereby eliminating the need for proxying
through the Kubernetes API server, which results in Internet network traffic and
associated costs.

It's important to note that the control plane Prometheus still requires the
proxy feature for scraping metrics from shoot nodes, as it operates within the
seed cluster and lacks direct network access to the shoot nodes.