WIP: Option 1: skip verification of the kubelet's certificate
Without this change, the Prometheus federation fails with the following
error in the local setup:

  Get "https://172.18.0.2:10250/metrics":
  tls: failed to verify certificate:
  x509: cannot validate certificate for 172.18.0.2
  because it doesn't contain any IP SANs

The reason for that is that the self issued certificate of the Kubelet:

  openssl s_client -connect 172.18.0.3:10250 -showcerts </dev/null 2>/dev/null | openssl x509 -noout -text | grep "Subject Alternative Name" -A 1
    X509v3 Subject Alternative Name:
      DNS:gardener-local-ha-multi-zone-control-plane

does not contain a SAN IP entry, unlike in a real cluster:

  openssl s_client -connect 10.254.74.24:10250 -showcerts </dev/null 2>/dev/null | openssl x509 -noout -text | grep "Subject Alternative Name" -A 1
    X509v3 Subject Alternative Name:
      DNS:ip-10-254-74-24.eu-west-1.compute.internal, IP Address:10.254.74.24

With this change, the PR works in the local setup but the certificate is not
verified.

Although it would be fine to skip certificate verification only in the local
setup (it is not straightforward to achieve it only for the local deployment),
we should anyways try to verify the certificate in real clusters to prevent man
in the middle attacks. Note that the impact of such an attack is fairly small
(kubelet metrics) but it is a valid (academic) goal to properly protect the
network path of the Kubelet scrape job.