Address review comment: Restrict the network policies
to the seed's node CIDR for the Kubelet network policy that was
introduced in this PR, and also for the node exporter network
policy which served as a blue print for the Kubelet network policy.
With this change, a minor TODO of the node exporter network policy
could be resolved.